requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
Markdown==3.4.1
pytz==2022.7
sqlparse==0.4.3
